Move runtime functions from model.ts to client.ts
Make model.ts a types-only module by moving wrapWorkerPort and
subscriptions to client.ts, their sole consumer. Imports in model.ts
are now type-only.
