+++ src/client.ts
@@ -1,13 +1,63 @@
 import * as Comlink from "comlink";
-import type { ObservableNotification, Observable } from "rxjs";
+import { Observable, Subject, type ObservableNotification } from "rxjs";
+import { dematerialize } from "rxjs/operators";
 import type { RegistryContract } from "./contract";
-import {
-  subscriptions,
-  wrapWorkerPort,
-  type Operations,
-  type SubscriptionKey,
-  type SubscriptionInput,
+import type {
+  Operations,
+  SubscriptionInput,
+  SubscriptionKey,
+  WorkerContract,
 } from "./model";
+
+/**
+ * Creates a Comlink remote proxy for the given worker contract using the provided MessagePort.
+ * @param port The MessagePort to communicate with the worker.
+ * @returns A Comlink remote proxy for the worker contract.
+ */
+export const wrapWorkerPort = <T extends Operations>(port: MessagePort) =>
+  Comlink.wrap<WorkerContract<T>>(port);
+
+export const subscriptions = <T extends Operations>(client: T) => {
+  function subscribe<K extends SubscriptionKey<T>>(
+    key: K,
+    ...args: SubscriptionInput<T, K> extends void
+      ? []
+      : [input: SubscriptionInput<T, K>]
+  ) {
+    type U = T[K] extends (
+      onNotification: (
+        notification: ObservableNotification<infer Update>,
+      ) => void,
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      ...args: any[]
+    ) => Promise<() => void>
+      ? Update
+      : never;
+
+    return new Observable<U>((subscriber) => {
+      const subscription = client[key] as unknown as (
+        onNotification: (notification: ObservableNotification<U>) => void,
+        ...args: SubscriptionInput<T, K> extends void
+          ? []
+          : [input: SubscriptionInput<T, K>]
+      ) => Promise<() => void>;
+
+      const notifications$ = new Subject<ObservableNotification<U>>();
+      const sub = notifications$.pipe(dematerialize()).subscribe(subscriber);
+      const onNotification = (n: ObservableNotification<U>) =>
+        notifications$.next(n);
+
+      const unsubscribePromise = subscription(onNotification, ...args);
+
+      return () => {
+        sub.unsubscribe();
+        notifications$.complete();
+        unsubscribePromise.then((f) => f());
+      };
+    });
+  }
+  return subscribe;
+};
 
 export interface CreateClientOptions {
   sharedWorker: SharedWorker;
